requirements.txt
Flask
pytesseract
Pillow
flask-cors
